Remove the old mechanism of file loading.
